# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "This is the API documentation for the Elonica project.",
        "description": "This is the API documentation for the Elonica project.",
        "contact": {
            "email": "pavlito97@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/messages/{id}/{offset}": {
            "get": {
                "tags": [
                    "messages"
                ],
                "description": "Возвращает историю переписки с другим пользователем.",
                "operationId": "App\\Http\\Api\\Controllers\\MessageController::chatHistory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Возвращает список сообщений."
                    },
                    "404": {
                        "description": "Сообщения не найдены."
                    }
                }
            }
        },
        "/messages/{id}": {
            "post": {
                "tags": [
                    "messages"
                ],
                "description": "Отправляет сообщение игроку.",
                "operationId": "App\\Http\\Api\\Controllers\\MessageController::sendMessage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "message",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "message",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Сообщение отправленно."
                    },
                    "422": {
                        "description": "Не удалось отправить сообщение."
                    }
                }
            },
            "delete": {
                "tags": [
                    "messages"
                ],
                "description": "Удаляет сообщение.",
                "operationId": "App\\Http\\Api\\Controllers\\MessageController::deleteMessage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Игрок успешно удален из списка друзей."
                    },
                    "404": {
                        "description": "Игрок не найден."
                    },
                    "422": {
                        "description": "Не удалось удалить сообщение."
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}